+++ src/main/java/com/authlete/mdoc/constants/MDLConstants.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright (C) 2023 Authlete, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.authlete.mdoc.constants;
+
+
+/**
+ * Constants for mDL.
+ *
+ * @since 1.12
+ *
+ * @see <a href="https://www.iso.org/standard/69084.html">ISO/IEC 18013-5:2021</a>
+ */
+public class MDLConstants
+{
+    /**
+     * The document type of the mDL
+     * ("{@code org.iso.18013.5.1.mDL}").
+     */
+    public static final String DOC_TYPE_MDL = "org.iso.18013.5.1.mDL";
+
+
+    /**
+     * The document type of the mDL version 1
+     * ("{@code org.iso.18013.5.1.mDL}").
+     */
+    public static final String DOC_TYPE_MDL_VER_1 = "org.iso.18013.5.1.mDL";
+
+
+    /**
+     * The name space of the MDL
+     * ("{@code org.iso.18013.5.1}").
+     */
+    public static final String NAME_SPACE_MDL = "org.iso.18013.5.1";
+
+
+    /**
+     * The name space of the MDL version 1
+     * ("{@code org.iso.18013.5.1}").
+     */
+    public static final String NAME_SPACE_MDL_VER_1 = "org.iso.18013.5.1";
+}